+++ components/breadcrumb/stories/breadcrumb-introduction.config.mjs
@@ -1,0 +1,14 @@
+export default {
+	sourcePath: '../readme.md',
+	outputName: 'breadcrumb-introduction',
+	story: {
+		title: 'Alpha/Components/Breadcrumb',
+		name: 'Introduction',
+		parameters: {
+			options: {
+				showPanel: false,
+				isToolshown: false
+			}
+		}
+	}
+};